+++ arculus-gcs/arculus-gcs-node/routes/taskRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const taskService = require('../services/taskService');
+const router = express.Router();
+
+router.get('/api/getTasks', taskService.getTasks);
+
+
+module.exports = router;
