+++ arculus-gcs/arculus-gcs-ui/src/config.js
@@ -1,4 +1,4 @@
 // config.js
-const API_URL = 'http://18.117.178.150:3001';
+const API_URL = 'http://18.119.101.183:3001';
 
 export { API_URL };
